+++ src/node/coinstats.cpp
@@ -6,6 +6,7 @@
 #include <node/coinstats.h>
 
 #include <coins.h>
+#include <crypto/muhash.h>
 #include <hash.h>
 #include <primitives/txid.h>
 #include <serialize.h>
@@ -41,6 +42,19 @@ static void ApplyHash(CCoinsStats &stats, CHashWriter &ss, const TxId &txid,
 static void ApplyHash(CCoinsStats &stats, std::nullptr_t, const TxId &txid,
                       const std::map<uint32_t, Coin> &outputs,
                       std::map<uint32_t, Coin>::const_iterator it) {}
+
+static void ApplyHash(CCoinsStats &stats, MuHash3072 &muhash, const TxId &txid,
+                      const std::map<uint32_t, Coin> &outputs,
+                      std::map<uint32_t, Coin>::const_iterator it) {
+    COutPoint outpoint = COutPoint(txid, it->first);
+    Coin coin = it->second;
+
+    CDataStream ss(SER_DISK, PROTOCOL_VERSION);
+    ss << outpoint;
+    ss << static_cast<uint32_t>(coin.GetHeight() * 2 + coin.IsCoinBase());
+    ss << coin.GetTxOut();
+    muhash.Insert(MakeUCharSpan(ss));
+}
 
 template <typename T>
 static void ApplyStats(CCoinsStats &stats, T &hash_obj, const TxId &txid,
@@ -109,6 +123,10 @@ bool GetUTXOStats(CCoinsView *view, CCoinsStats &stats,
             CHashWriter ss(SER_GETHASH, PROTOCOL_VERSION);
             return GetUTXOStats(view, stats, ss, interruption_point);
         }
+        case (CoinStatsHashType::MUHASH): {
+            MuHash3072 muhash;
+            return GetUTXOStats(view, stats, muhash, interruption_point);
+        }
         case (CoinStatsHashType::NONE): {
             return GetUTXOStats(view, stats, nullptr, interruption_point);
         }
@@ -120,9 +138,16 @@ bool GetUTXOStats(CCoinsView *view, CCoinsStats &stats,
 static void PrepareHash(CHashWriter &ss, const CCoinsStats &stats) {
     ss << stats.hashBlock;
 }
+// MuHash does not need the prepare step
+static void PrepareHash(MuHash3072 &muhash, CCoinsStats &stats) {}
 static void PrepareHash(std::nullptr_t, CCoinsStats &stats) {}
 
 static void FinalizeHash(CHashWriter &ss, CCoinsStats &stats) {
     stats.hashSerialized = ss.GetHash();
+}
+static void FinalizeHash(MuHash3072 &muhash, CCoinsStats &stats) {
+    uint256 out;
+    muhash.Finalize(out);
+    stats.hashSerialized = out;
 }
 static void FinalizeHash(std::nullptr_t, CCoinsStats &stats) {}