+++ doc/release-notes.md
@@ -5,3 +5,5 @@ Bitcoin ABC version 0.25.2 is now available from:
   <https://download.bitcoinabc.org/0.25.2/>
 
 This release includes the following features and fixes:
+ - The `gettxoutsetinfo` RPC now accepts `'muhash'` as a value for the `hash_type`
+   input parameter, in addition to `'none'` and `'hash_serialized'`.